requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.18
django-allauth==0.54.0
django-appconf==1.0.5
django-ckeditor==6.5.1
django-classy-tags==4.0.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-imagekit==4.1.0
django-js-asset==2.0.0
django-recaptcha==3.0.0
django-scheduler==0.10.1
django-taggit==4.0.0
django-taggit-templatetags==0.2.5
django-taggit-templatetags2==1.6.1
django-templatetag-sugar==1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
frozenlist==1.3.3
geographiclib==2.0
geopy==2.3.0
gpxpy==1.5.0
icalendar==5.0.5
idna==3.4
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
multidict==6.0.4
mysql-connector-python==8.0.33
oauthlib==3.2.2
parso==0.8.3
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psycopg2==2.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.6
xmltodict==0.13.0
yarl==1.9.2
